modefied build-js

+++ build-script.js
@@ -35,26 +35,31 @@ esbuild
     },
   })
   .then(() => {
-    //1 Copy swagger.json after successful build
+    // 1. Copy swagger.json after successful build
     fs.copySync(
       path.resolve(__dirname, "src/docs/swagger.json"),
       path.resolve(__dirname, "build/docs/swagger.json")
     );
     console.log("Swagger JSON copied successfully!");
 
-    //2 Copy ecosystem.config.js
+    // 2. Copy ecosystem.config.js
     fs.copySync(
       path.resolve(__dirname, "ecosystem.config.js"),
       path.resolve(__dirname, "build/ecosystem.config.js")
     );
     console.log("Ecosystem Config copied successfully!");
 
-    //3 Copy ecosystem.config.js
-    fs.copySync(
-      path.resolve(__dirname, "src/configs/.env.development"),
-      path.resolve(__dirname, "build/configs/.env.local")
-    );
-    console.log("Ecosystem Config copied successfully!");
+    // 3. Check if .env.development exists before copying
+    const envPath = path.resolve(__dirname, "src/configs/.env.development");
+    if (fs.existsSync(envPath)) {
+      fs.copySync(
+        envPath,
+        path.resolve(__dirname, "build/configs/.env.local")
+      );
+      console.log(".env.development copied successfully!");
+    } else {
+      console.warn(".env.development file not found, skipping copy.");
+    }
   })
   .catch((error) => {
     console.error("Build failed:", error);
